Revert "Refactored and added picker"
This reverts commit 0cdff863f6cb387fc724c25b9d20b5455cda27bf.
