+++ js/containers/Root.js
@@ -4,12 +4,12 @@
 import React, { Component } from 'react'
 import { Provider } from 'react-redux'
 import configureStore from '../configureStore'
-import Tindspect from '../components/tindspectHolder';
+import Tindspect from '../components/overallComponents';
 
 
 const store = configureStore()
  
-export default class PickerRoot extends Component {
+export default class Root extends Component {
     render() {
         return (
             <Provider store={store}>
